refactor: move updating of categories when venue changes into useCurrentVenue hook, since it is a side effect of venue changes.
