+++ packages/map-template/src/atoms/appConfigState.js
@@ -1,0 +1,8 @@
+import { atom } from 'recoil';
+
+const appConfigState = atom({
+    key: 'appConfig',
+    default: undefined
+});
+
+export default appConfigState;
